moved the trim function to a single module
